+++ src/app/f/[folderId]/page.tsx
@@ -2,7 +2,7 @@ import { z } from "zod";
 
 import * as queries from "~/server/db/queries";
 
-import DriveContents from "../../drive-contents";
+import DriveContents from "../drive-contents";
 
 export default async function GoogleDriveClone(props: {
   params: Promise<{ folderId: number }>;
